Added the "SAMs can pretend to log in as their cutomers" feature (you can choose to filter out your group memberships using a dropdown list on the dashboard summary and kpi view screens).
Added a "sticky groups" user pref that helps users not lose
permissions on their KPIs when using the new group filter.  For
example, Warren Shepell SAMs will need DASHBOARD_SAM as a sticky group
to continue using the CUSTOMER_* KPIs as a customer.

Updated the dimension filter tree to only show dimension values that
exist in the status table for the current date range.