Fix for Bug 1367
The problem was that the SPDataSourceType that is used to get the JDBC url template was not getting set in the PlatformSpecificConnectionOptionPanel if the data source type was set for that connection.

Additionally, the 'updateFieldsFromURL' method in the PlatformSpecificConnectionOptionPanel  was not dealing with null templates properly, so this was still a problem if editing a connection's JDBC URL before selecting a data source type.