Added optional 'javadoc' and 'javadoc.with.umlgraph' build target to generate javadoc.
The 'with.umlgraph' target generates them with UML diagrams generated with the UMLGraph library. The one prerequisite is that you must have Graphviz installed and the 'dot' executable on your PATH before running this. Also, running this on Windows XP doesn't seem to work for some reason.

It's also important to note that the UmlGraph JAR being committed is version 4.8, which is the last version that is compatible with Java 5 (all future versions require Java 6).