Backport of r733 and r734 into the architect-0.9.12 branch.
This fixes an issue where the Table Copy feature would throw an exception when copying data to a MySQL table. This was because the Table Copy feature did not set autocommit to false when running the table copy. This also meant that the table copy was committing all its commands instead of waiting for the entire table copy transaction to finish before committing.

Oddly enough, PostgreSQL, SQL Server 2005, and Oracle 10g Express did not throw exceptions when committing while automcommit was still true.

Also changed System.out and System.err calls to the appropriate Log4j logger call and added a missing BSD license notice.