The SQLQueryUIComponents no longer allows displaying only the first result table. The ability to display only one tab was used in Wabit but is no longer needed as we want to show the log.
The log file is now appended to instead of replacing the last 
executed log. The log also has a timestamp and length of execution. 
The log can also be saved to a file.

The search field affects all results at the same time. This allows
one search field to be made that affects all results in Wabit.