Removed much of the passing-around of DatabaseMetaData that was added least year in an attempt to speed up reverse engineering. That stuff ended up not being used, but it complicated the populate API, which we are now overhauling to support refresh.
Also implemented a simple refresh method on SQLObject that specifies how a refresh is initiated, but it's currently just a skeleton.

There is also a new test class, RefreshTest, where the tests for the refresh feature will live. The sole test in there is currently failing because refresh is not yet implemented.