Classpath cleanup:
* Removed duplicate servlet API jar file
* Removed Struts jar, which wasn't being used at all
* Removed HSQLDB from classpath (it's supposed to be loaded via the SPDataSource system)

Bug Fix (related to HSQLDB change): now JDBCClassLoader instances are cached and shared across SPDataSourceType instances. The cache key is an immutable list of the jar files in the classloader's classpath. This fix makes it possible to run the test suite without running out of the default PermGen memory allocation.