Pulled out an old bit of code that was trying to compensate for DBMD.getColumns() returning values that were either too high, or 0.
This kind of stuff should be in the JDBC wrappers now, so I've removed it from SQLColumn. It turns out that HSQLDB reports integer columns as having a size (precision) of 0, so I implemented the fix in our HSQLDB wrapper.

There will likely be more fixes that have to be added to the wrappers as we discover them.