Set the Maven output directory to match the Eclipse output directory.
It seems that this confusion may be what was causing dependent projects not picking up source code changes in the library, even though they were configured to use the workspace projects.

Also added back a commented out option for the compiler that got removed for some reason. If you want to use the pom file with a non-default JVM, you would need that fork tag.