NEW - bug 1866: Add delete functionality for OLAP Queries http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1866
Adds the ability to delete OLAP queries via the right-click menu. Also adds the functionality to remove OLAP queries related to an OLAP datasource if that OLAP datasource is removed from the workspace (almost said project there ;P). One issue remaining is that the 'Create New' button in the dialog that appears if you decide to replace an OLAP datasource that you're deleting will open up a JDBC data source creation panel instead of an OLAP one at the moment. The Datasource user prompter dialog assumes that the new data source would be a JDBC data source.