REOPENED - bug 2020: A warning before executing a query with a missing join in the querypen http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2020
Fixed the prompt when the director table is added to the 'highly rated
actors' table. The containsCrossJoins method is updated to confirm that
each table is connected to one of the tables that came before it 
instead of needing to be connected to the table that comes directly
before it.