This commit is to support Wabit's addChild() method on WabitObjects.
The Items contained in the QueryImpl that tracked the selection and
order by ordering have been removed and the order of the Items is 
stored in the items themselves. This ends up reducing complexity by
removing the wrappers on Items so they can be added in the query and
the further wrappers in Wabit so the wrappers are wrapped by proper
WabitObjects. The Query has methods for selecting and ordering items
to properly keep the item's ordering values valid.

Several new methods have been added to the Query interface to allow
adding children to a specific index.