NEW - bug 2360: Copying a SQL statement into a query causes Wabit client to crash http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2360
The root cause of this issue was the event for changing the user modified query text was not firing when switching between using a generated query from the GUI and the text modified by the user. Now an event will be fired on any change of the user modified query and an event will be fired when the user's text modifications are removed.