Adding google-collections jar to the library. Using Multimap to keep track of which exceptions Mutator annotated methods throw, to generate try catch blocks accordingly in persister helpers.
Added the Mutator and Accessor annotations for the parent property. Also changed the ConstructorParameter annotaton "propertyName" field to "value" so that any arguments passed into the annotation can be done as default (without 'propertyName = ' prefixed).

The formatting of the generated persister helpers are a bit better now. There is line spacing between methods, and imports are alphabetized.