Annotated the Grant, Group, GroupMember and User classes.
Also fixed the commitProperty and findProperty methods in the SPAnnotationProcessorFactory so that they do not take a SessionPersisterSuperConverter as a parameter because there is already a converter stored as a field when the constructor is called.

Additionally, fixed the generation of the commitObject method in the persister helpers where Character.valueOf should take a char value instead of a string.