Fixed how the SPAnnotationProcessor outputs source files. Instead of having the processor specify to output files to the "src" directory, it is passed as an argument in the build.xml apt task. The processor only needs to worry about package names now, it does not need to know about where the source directory is.
Everything in build.xml that used to reference the "compile" target has now been changed to compile-with-annotations (previously called process-annotations).

Also when the branch was merged into the trunk with the SPObject model change, somehow the pom.xml file wasn't changed to include the apt-maven-plugin. Further changes need to be made to the POM so that it generates the persister helpers and SPPersisterHelperFactoryImpl to the right package and compiles correctly.