RESOLVED - bug 2702: Ensure populate still works for all possible relationship orderings http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2702
This was actually a problem with our Oracle JDBC wrapper's getExportedKeys method--it was ordering by the wrong set of columns.

Also, there was some doubt over whether or not the FK table was being set up properly in column mappings. It was, but this led me to discover that the getShortDisplayName() method was printing out the wrong thing (should have been pkCol - fkTable.fkCol)