NEW - bug 2697: Remove the SPPersisterHelperFactoryImpl http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2697
Removed the SPPersisterHelperFactory and the implementation made by the annotation factory. The PersisterHelperFinder contains utility methods to access persister helpers instead of the factory. The persister helpers are also located under a 'generated' package in the package that the object to be persisted exists in. The reason for removing the SPPersisterHelperFactory is it was limiting the use of the annotation factory to only the library. We now need persister helpers in Architect so classes need a way to access the helpers outside of the library.

There is a test in the PersistedSPObjectTest for rollback that is commented out. The issue of persisters needing to exist in Architect is blocking other work.