Refactored a Popup creator from Wabit into the library. This popup was used in Wabit to choose cubes in an OLAP query. It has been moved into the library to allow the ColumnEditPanel in Architect to use a popup for choosing domains and data types.
If two or more users are modifying the same column type and one user commits the change, the user who has not committed the change will see that their domain/type tree will be highlighted in red or yellow, just as the other components in the panel do.

Also, the name property in UserDefinedSQLType is now a constructor property. This change will allow the Architect play pen to use the correct name on load.