NEW - bug 2833: Dragging in multiple columns into the PK of a parent table adds them in reverse order to the FK table http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2833
Changed the decision on the index location of the column being added to the FK table. When adding multiple columns to the 
primary key index of the PK table it is not guaranteed that they will be added from first to last, some places iterate 
backwards. To handle this easier for the future the location to insert the column in the FK table is decided on by the
columns that exist in the PK and FK table and their current order.